make nexus deployable to now (#517)
* fix(build): require package.json and plugins

* feat: add graphql path option

* fix: find package.json decoupled from cwd

* refactor: constants and compile

* fix: decouple layout project root from cwd if possible

* fix(plugins): return no plugin if no package.json

+++ src/lib/tsc.ts
@@ -152,14 +152,23 @@ export function getTSIncrementalFilePath(layout: Layout) {
  */
 export function compile(
   program: ts.EmitAndSemanticDiagnosticsBuilderProgram,
-  layout: Layout
+  layout: Layout,
+  opts?: { skipTSErrors?: boolean; removePreviousBuild?: boolean }
 ): void {
-  log.trace('remove previous build folder if present...')
-  fs.remove(layout.buildOutput)
+  if (opts?.removePreviousBuild === true) {
+    log.trace('remove previous build folder if present...')
+    fs.remove(layout.buildOutput)
+  }
+
   log.trace('done')
   log.trace('emit transpiled modules to disk...')
   const emitResult = program.emit()
   log.trace('done', { filesEmitted: emitResult.emittedFiles?.length ?? 0 })
+
+  if (opts?.skipTSErrors === true) {
+    return
+  }
+
   const allDiagnostics = ts
     .getPreEmitDiagnostics(program.getProgram())
     .concat(emitResult.diagnostics)
